Temporarily rename SWRConfigClient
